+++ src/main/resources/site/parts/timeline/timeline.ts
@@ -3,7 +3,14 @@ import { localize } from '/lib/xp/i18n'
 import { render } from '/lib/enonic/react4xp'
 import { renderError } from '/lib/ssb/error/error'
 import { imageUrl, getImageAlt } from '/lib/ssb/utils/imageUtils'
-import { type TimelineElement, type TimelineEvent } from '/lib/types/partTypes/timeline'
+import {
+  type Event,
+  type SimpleBox,
+  type ExpansionBox,
+  type DirectorBox,
+  type TimelineElement,
+  type TimelineEvent,
+} from '/lib/types/partTypes/timeline'
 import { forceArray } from '/lib/ssb/utils/arrayUtils'
 import { type Timeline as TimelinePartConfig } from '/site/parts/timeline'
 
@@ -24,10 +31,13 @@ function renderPart(req: XP.Request) {
 
   const timelineConfig: TimelinePartConfig = part.config
   const timelineItems: TimelinePartConfig['TimelineItemSet'] = forceArray(timelineConfig.TimelineItemSet)
+
   const timelineProps: TimelineElement[] = timelineItems.map((element) => {
+    const events: Event[] = element.event ? forceArray(element.event) : []
+    const parsedEvents: TimelineEvent[] = parseEvents(events)
     return {
       year: element.year,
-      event: element.event ? parseEvent(forceArray(element.event)) : [],
+      event: parsedEvents,
     }
   })
 
@@ -47,37 +57,86 @@ function renderPart(req: XP.Request) {
   })
 }
 
-function parseEvent(events: TimelineEvent[]): TimelineEvent[] {
-  const parsedElements = events.map((event) => {
-    const image = event.directorImage
+function parseEvents(events: Event[]): TimelineEvent[] {
+  return events.map((event: Event) => {
+    return parseEvent(event)
+  })
+}
+
+function parseEvent(event: Event): TimelineEvent {
+  if (event.simpleBox) {
+    return parseSimpleBox(event.simpleBox)
+  }
+
+  if (event.expansionBox) {
+    return parseExpansionBox(event.expansionBox)
+  }
+  if (event.directorBox) {
+    return parseDirector(event.directorBox)
+  }
+  return {
+    eventType: '',
+    title: '',
+    text: '',
+    directorImage: undefined,
+    directorImageAltText: '',
+    timelineCategory: '',
+    targetUrl: '',
+  }
+}
+
+function parseSimpleBox(event: SimpleBox): TimelineEvent {
+  const simpleBox = {
+    eventType: 'simpleBox',
+    title: event.title,
+    text: event.text ?? '',
+    directorImage: undefined,
+    directorImageAltText: '',
+    timelineCategory: event.timelineCategory,
+    targetUrl: event.urlContentSelector ? getLinkTargetUrl(event) : '',
+  }
+  return simpleBox
+}
+
+function parseExpansionBox(event: ExpansionBox): TimelineEvent {
+  return {
+    eventType: 'expansionBox',
+    title: event.title,
+    text: event.text,
+    directorImage: undefined,
+    directorImageAltText: '',
+    timelineCategory: event.timelineCategory,
+    targetUrl: '',
+  }
+}
+
+function parseDirector(event: DirectorBox): TimelineEvent {
+  return {
+    eventType: 'directorBox',
+    title: event.title,
+    text: event.text,
+    directorImage: event.directorImage
       ? imageUrl({
           id: event.directorImage as string,
           scale: 'block(100,100)',
           format: 'jpg',
         })
-      : undefined
-
-    const imageAltText = event.directorImage ? getImageAlt(event.directorImage) : ''
-    return {
-      ...event,
-      directorImage: image,
-      directorImageAltText: imageAltText ?? '',
-      targetUrl: getLinkTargetUrl(event),
-    }
-  })
-
-  return parsedElements
+      : undefined,
+    directorImageAltText: event.directorImage ? getImageAlt(event.directorImage) : '',
+    timelineCategory: 'director',
+    targetUrl: event.urlContentSelector ? getLinkTargetUrl(event) : '',
+  }
 }
 
-function getLinkTargetUrl(event: TimelineEvent): string {
-  if (event.urlArticle?._selected == 'optionLink') {
-    return event.urlArticle.optionLink.link ?? ''
+function getLinkTargetUrl(event: SimpleBox | DirectorBox): string {
+  if (event.urlContentSelector?._selected == 'optionLink') {
+    return event.urlContentSelector.optionLink.link ?? ''
   }
 
-  if (event.urlArticle?._selected == 'optionXPContent') {
-    return event.urlArticle.optionXPContent.xpContent
+  if (event.urlContentSelector?._selected == 'optionXPContent') {
+    return event.urlContentSelector.optionXPContent.xpContent
       ? pageUrl({
-          id: event.urlArticle.optionXPContent.xpContent,
+          id: event.urlContentSelector.optionXPContent.xpContent,
         })
       : ''
   }
