+++ src/main/resources/site/parts/timeline/timeline.xml
@@ -9,71 +9,66 @@
                     <label>År</label>
                     <occurrences minimum="1" maximum="1"/>  
                 </input>
-
-                <item-set name="event">  
-                    <label>Hendelser</label>  
-                    <occurrences minimum="0" maximum="0"/>  
-                    <items>
-                        <input name="title" type="TextLine">
-                            <label>Tittel</label>
-                            <occurrences minimum="1" maximum="1"/>  
-                        </input>
-                        <input name="ingress" type="TextLine">
-                            <label>Ingress</label>
-                            <occurrences minimum="0" maximum="1"/>  
-                        </input>
-                        <input name="eventText" type="HtmlArea">
-                            <label>Tekst</label>
-                            <config>
-                                <include>Bold Italic Underline Strike Subscript Superscript Cut Copy Blockquote</include>
-                                <allowHeadings>h2 h3 h4 h5</allowHeadings>
-                            </config>
-                        </input>
-                        <input name="directorImage" type="ImageSelector">
-                            <label>Direktørbilde</label>
-                            <occurrences minimum="0" maximum="1"/>
-                        </input>
-                        <input name="category" type="RadioButton">
-                            <label>Hendelsestype</label>
-                            <occurrences minimum="1" maximum="1"/>
-                            <config>
-                                <option value="statistic">Statistikk</option>
-                                <option value="aboutSsb">Om SSB</option>
-                                <option value="director">Direktør</option>
-                            </config>
-                            <default>aboutSsb</default>
-                        </input>
-                        <option-set name="urlArticle">
-                            <label>Artikkel</label>
-                            <expanded>false</expanded>
-                            <occurrences minimum="0" maximum="1"/>
-                            <help-text>Velge enten url eller innhold </help-text>
-                            <options minimum="0" maximum="1">
-                                <option name="optionLink">
-                                    <label>URL</label>
-                                    <items>
-                                        <input name="link" type="TextLine">
-                                            <label>Lenke</label>
-                                            <occurrences minimum="0" maximum="1"/>
-                                        </input>
-                                    </items>
-                                </option>
-                                <option name="optionXPContent">
-                                    <label>XP-innhold</label>
-                                    <items>
-                                    <input name="xpContent" type="ContentSelector">
-                                        <label>Innhold i XP</label>
-                                        <occurrences minimum="0" maximum="1"/>
-                                        <config>
-                                            <allowPath>${site}</allowPath>
-                                        </config>
-                                    </input>
-                                    </items>
-                                </option>
-                            </options>
-                        </option-set>
-                    </items>
-                </item-set>
+                <option-set name="event">
+                    <label>Hendelse</label>
+                    <expanded>false</expanded>
+                    <occurrences minimum="0" maximum="0"/>
+                    <help-text>Velg type hendelse</help-text>
+                    <options minimum="0" maximum="1">
+                        <option name="simpleBox">
+                            <label>Enkel tekstboks med eller uten lenke</label>
+                            <help-text>Brukes hvis man kun har en tittel og eventuelt litt tekst</help-text>
+                            <items>
+                                <input name="title" type="TextLine">
+                                    <label>Tittel</label>
+                                    <occurrences minimum="1" maximum="1"/>
+                                </input>
+                                <input name="text" type="TextLine">
+                                    <label>Tekst</label>
+                                    <occurrences minimum="0" maximum="1"/>  
+                                </input>
+                                <mixin name="linkUrlOrContent" />
+                                <mixin name="timelineCategory" />
+                            </items>
+                        </option>
+                        <option name="expansionBox">
+                            <label>Utvidbar box (ExpansionBox)</label>
+                            <help-text>Brukes hvis man ønsker å ha en lengre tekst som ikke lenker til artikkel</help-text>
+                            <items>
+                                <input name="title" type="TextLine">
+                                    <label>Tittel</label>
+                                    <occurrences minimum="1" maximum="1"/>
+                                </input>
+                                <input name="text" type="HtmlArea">
+                                    <label>Tekst</label>
+                                    <config>
+                                        <include>Bold Italic Underline Strike Subscript Superscript Cut Copy Blockquote</include>
+                                        <allowHeadings>h2 h3 h4 h5</allowHeadings>
+                                    </config>
+                                </input>
+                                <mixin name="timelineCategory" />
+                            </items>
+                        </option>
+                        <option name="directorBox">
+                            <label>Direktør</label>
+                            <items>
+                                <input name="title" type="TextLine">
+                                    <label>Navn</label>
+                                    <occurrences minimum="1" maximum="1"/>
+                                </input>
+                                <input name="text" type="TextLine">
+                                    <label>Tekst</label>
+                                    <occurrences minimum="0" maximum="1"/>  
+                                </input>
+                                <input name="directorImage" type="ImageSelector">
+                                    <label>Direktørbilde</label>
+                                    <occurrences minimum="0" maximum="1"/>
+                                </input>
+                                <mixin name="linkUrlOrContent" />
+                            </items>
+                        </option>
+                    </options>
+                </option-set>
             </items>
         </item-set>
   </form>
